Add comment/reply system to Activity Log sidebar
- Integrate comments from mission, goals, and objectives into Activity Log
- Implement threaded reply display with nested layout
- Add reply dialog with parent comment context and entity info
- Filter to show only parent comments in main log (replies nested)
- Add visual distinction for comments (primaryLight badge with icon)
- Display replies indented with left border and distinct styling
- Move author to same line as timestamp, right-justified
- Add debug logging for comment queries and reply loading
- Update _LogEntryData to include commentId, entityId, and originalEntityType

+++ lib/core/models/question.dart
@@ -51,10 +51,10 @@ class Question {
   final bool? allowMultipleSelections;
 
   /// Minimum value for scale questions (null for other types)
-  final int? scaleMin;
+  final double? scaleMin;
 
   /// Maximum value for scale questions (null for other types)
-  final int? scaleMax;
+  final double? scaleMax;
 
   /// Labels for scale endpoints (e.g., ["Not at all", "Extremely"])
   final List<String>? scaleLabels;
@@ -124,8 +124,8 @@ class Question {
     QuestionType? questionType,
     List<String>? options,
     bool? allowMultipleSelections,
-    int? scaleMin,
-    int? scaleMax,
+    double? scaleMin,
+    double? scaleMax,
     List<String>? scaleLabels,
     int? answerCharacterLimit,
     int? order,
